Rename worker module
